wip: some new button tests
