+++ src/Templates/Styles/button.css
@@ -1,6 +1,10 @@
+.btn {
+    padding: 10px;
+}
+
 .button {
     display: inline-block;
-    margin: 0px 10px 0px 10px;
+    margin: 0px 0px 0px 10px;
     padding: 10px;
     text-align: center;
     text-decoration: none;
